% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econio-package.R
\docType{package}
\name{econio-package}
\alias{econio}
\alias{econio-package}
\title{econio: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
