% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{io_check_axes}
\alias{io_check_axes}
\title{Check input-output table axes}
\usage{
io_check_axes(data)
}
\arguments{
\item{data}{An input-output table.}
}
\value{
An \code{econ_io_table} object.
}
\description{
Check input-output table axes
}
