% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{io_check_totals}
\alias{io_check_totals}
\title{Check input-output table totals}
\usage{
io_check_totals(data, total_tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{data}{An input-output table.}

\item{total_tolerance}{A numeric. The tolerance for the total check. By
default, \code{.Machine$double.eps^0.5}.}
}
\value{
An \code{econ_io_table} object.
}
\description{
Check input-output table totals
}
