% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.R
\name{io_ghosh_inverse}
\alias{io_ghosh_inverse}
\title{Ghosh inverse matrix}
\usage{
io_ghosh_inverse(data, open_economy = NULL)
}
\arguments{
\item{data}{An \code{econ_io_table} object.}

\item{open_economy}{A scalar logical. If \code{TRUE}, open economy assumptions are
used.}
}
\value{
An \code{econ_io_table} object of the Ghosh inverse matrix.
}
\description{
Ghosh inverse matrix
}
