% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{io_reclass}
\alias{io_reclass}
\title{Reclassify an input-output table}
\usage{
io_reclass(
  data,
  input_region_data = NULL,
  input_sector_data = NULL,
  output_region_data = NULL,
  output_sector_data = NULL,
  from_col = "from",
  to_col = "to",
  weight_col = "weight",
  weight_tolerance = .Machine$double.eps^0.5,
  check_axes = TRUE
)
}
\arguments{
\item{data}{An input-output table.}

\item{input_region_data, input_sector_data, output_region_data, output_sector_data}{Data frames used to reclassify the input and output regions and sectors.
If \code{NULL}, no reclassification is performed for that axis. By default,
\code{NULL}.}

\item{from_col, to_col, weight_col}{Names of the columns in the
reclassification data. Defaults are \code{"from"}, \code{"to"}, and \code{"weight"}.}

\item{weight_tolerance}{Tolerance for checking that weights sum to 1.
By default, \code{.Machine$double.eps^0.5}.}

\item{check_axes}{A scalar logical. If \code{TRUE}, the input and output axes are
checked to be identical. By default, \code{TRUE}.}
}
\value{
A reclassified input-output table.
}
\description{
Reclassify an input-output table
}
