% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{io_sector_name}
\alias{io_sector_name}
\title{Get sector names}
\usage{
io_sector_name(x)
}
\arguments{
\item{x}{An \code{econ_io_sector} object or a data frame with a column named
\code{sector}.}
}
\value{
A vector of sector names.
}
\description{
Get sector names
}
