% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{io_sector_type}
\alias{io_sector_type}
\title{Get sector types}
\usage{
io_sector_type(x)
}
\arguments{
\item{x}{An \code{econ_io_sector} object or a data frame with a column named
\code{sector}.}
}
\value{
A vector of sector types.
}
\description{
Get sector types
}
