% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-to.R
\name{io_table_to_byproduct_transfer}
\alias{io_table_to_byproduct_transfer}
\title{Convert byproduct transactions from the Stone method to the transfer method}
\usage{
io_table_to_byproduct_transfer(data)
}
\arguments{
\item{data}{An input-output table.}
}
\value{
An input-output table with negative domestic intermediate
transactions transferred to the corresponding reverse-direction cell.
}
\description{
Apply this after {\link{io_table_to_noncompetitive_import}} if \code{data} is a
competitive import type table, since that conversion can itself introduce
negative domestic intermediate transactions that this function would not
yet see.
}
