% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-to.R
\name{io_table_to_regional}
\alias{io_table_to_regional}
\title{Convert a multiregional input-output table to a regional input-output table}
\usage{
io_table_to_regional(data, export_sector_name, import_sector_name)
}
\arguments{
\item{data}{A multiregional input-output table}

\item{export_sector_name}{A name for the export sector.}

\item{import_sector_name}{A name for the import sector.}
}
\value{
A list of regional input-output tables.
}
\description{
Convert a multiregional input-output table to a regional input-output table
}
