% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{io_total_output}
\alias{io_total_output}
\title{Get total output}
\usage{
io_total_output(
  data,
  input_sector_type = "industry",
  output_sector_type = c("industry", "final_demand", "export", "import"),
  same_region = FALSE
)
}
\arguments{
\item{data}{An \code{econ_io_table} object.}

\item{input_sector_type}{A character vector of input sector types.}

\item{output_sector_type}{A character vector of output sector types.}

\item{same_region}{A logical scalar. If \code{TRUE}, values between different
regions are ignored.}
}
\value{
An \code{econ_io_table} object of total output.
}
\description{
Get total output
}
