% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econio-package.R
\docType{package}
\name{econio-package}
\alias{econio}
\alias{econio-package}
\title{econio: Input-Output Analysis}
\description{
Provides a set of functions for input-output analysis.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
