% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_coef.R
\name{io_import_coef}
\alias{io_import_coef}
\title{Import coefficients}
\usage{
io_import_coef(data)
}
\arguments{
\item{data}{An \code{econ_io_table} object.}
}
\value{
An \code{econ_io_table} object of import coefficients.
}
\description{
Import coefficients
}
