% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief_inverse.R
\name{io_leontief_inverse}
\alias{io_leontief_inverse}
\title{Leontief inverse matrix}
\usage{
io_leontief_inverse(data, endogenize_import = NULL)
}
\arguments{
\item{data}{An \code{econ_io_table} object.}

\item{endogenize_import}{A scalar logical. If \code{TRUE}, imports are
endogenized.}
}
\value{
An \code{econ_io_table} object of the Leontief inverse matrix.
}
\description{
Leontief inverse matrix
}
