% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{io_output_coef}
\alias{io_output_coef}
\title{Output coefficients}
\usage{
io_output_coef(data, open_economy = NULL)
}
\arguments{
\item{data}{An \code{econ_io_table} object.}

\item{open_economy}{A scalar logical. If \code{TRUE}, open economy assumptions are
used.}
}
\value{
An \code{econ_io_table} object of output coefficients.
}
\description{
Output coefficients
}
