% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{io_table_multiregional}
\alias{io_table_multiregional}
\title{Create a multi-regional input-output table}
\usage{
io_table_multiregional(
  data,
  input = c("input_region", "input_sector_type", "input_sector_name"),
  output = c("output_region", "output_sector_type", "output_sector_name"),
  competitive_import = NULL,
  total_tolerance = .Machine$double.eps^0.5
)
}
\arguments{
\item{data}{A data frame.}

\item{input}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Input region, sector type
and name columns.}

\item{output}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Output region, sector type
and name columns.}

\item{competitive_import}{A scalar logical. If \code{TRUE}, the table is assumed
to be a competitive import type.}

\item{total_tolerance}{A numeric. The tolerance for the total check. By
default, \code{.Machine$double.eps^0.5}.}
}
\value{
An \code{econ_io_table} object.
}
\description{
Create a multi-regional input-output table
}
