% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{io_table_regional}
\alias{io_table_regional}
\title{Create a regional input-output table}
\usage{
io_table_regional(
  data,
  input_cols = c("input_sector_type", "input_sector_name"),
  output_cols = c("output_sector_type", "output_sector_name"),
  competitive_import = NULL,
  total_tolerance = .Machine$double.eps^0.5,
  check_axis = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{input_cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Input sector type and name
columns.}

\item{output_cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Output sector type and
name columns.}

\item{competitive_import}{A scalar logical. If \code{TRUE}, the table is assumed
to be a competitive import type.}

\item{total_tolerance}{A numeric. The tolerance for the total check. By
default, \code{.Machine$double.eps^0.5}.}

\item{check_axis}{A scalar logical. If \code{TRUE}, the input and output axes are
checked to be identical. Default is \code{TRUE}.}
}
\value{
An \code{econ_io_table} object.
}
\description{
Create a regional input-output table
}
