% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{io_table_regional}
\alias{io_table_regional}
\title{Create a regional input-output table}
\usage{
io_table_regional(
  data,
  input = c("input_sector_type", "input_sector_name"),
  output = c("output_sector_type", "output_sector_name"),
  competitive_import = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{input}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Input sector type and name
columns.}

\item{output}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Output sector type and
name columns.}

\item{competitive_import}{A scalar logical. If \code{TRUE}, the table is assumed
to be a competitive import type.}
}
\value{
An \code{econ_io_table} object.
}
\description{
Create a regional input-output table
}
