% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-to.R
\name{io_table_to_competitive_import}
\alias{io_table_to_competitive_import}
\title{Convert an input-output table to a competitive import type table}
\usage{
io_table_to_competitive_import(data)
}
\arguments{
\item{data}{An input-output table}
}
\value{
A competitive import type input-output table
}
\description{
Convert an input-output table to a competitive import type table
}
