% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-to.R
\name{io_table_to_noncompetitive_import}
\alias{io_table_to_noncompetitive_import}
\title{Convert an input-output table to a noncompetitive import type table}
\usage{
io_table_to_noncompetitive_import(data, import_sector_name = NA_character_)
}
\arguments{
\item{data}{An input-output table}

\item{import_sector_name}{A name for the import sector. By default,
\code{NA_character_}.}
}
\value{
A noncompetitive import type input-output table.
}
\description{
Convert an input-output table to a noncompetitive import type table
}
