% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{io_total_input}
\alias{io_total_input}
\title{Get total input}
\usage{
io_total_input(
  data,
  input_sector_type = c("industry", "import", "value_added"),
  same_region = FALSE
)
}
\arguments{
\item{data}{An \code{econ_io_table} object.}

\item{input_sector_type}{A character vector of input sector types.}

\item{same_region}{A logical scalar. If \code{TRUE}, values between different
regions are ignored.}
}
\value{
An \code{econ_io_table} object of total input.
}
\description{
Get total input
}
